Fixed up some compile issues. Changed type for MaxBufferSize to size_t. Unsigned long was not appropriate.
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/sip/resiprocate@177 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ UdpTransport.hxx
@@ -19,7 +19,7 @@ class UdpTransport : public Transport
       virtual void process() ;
 
    private:
-      static const unsigned long MaxBufferSize;
+      static const size_t MaxBufferSize;
 };
  
 }
